// frontend/src/types/pagination.ts
export interface PageResponse<T> {
    content: T[]
    pageable: {
      pageNumber: number
      pageSize: number
      sort: {
        empty: boolean
        sorted: boolean
        unsorted: boolean
      }
      offset: number
      unpaged: boolean
      paged: boolean
    }
    last: boolean
    totalElements: number
    totalPages: number
    size: number
    number: number
    sort: {
      empty: boolean
      sorted: boolean
      unsorted: boolean
    }
    first: boolean
    numberOfElements: number
    empty: boolean
  }
  